+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 -   Add start duration in result of start.
 -   Use argumentsToStringReduced for internal dev logs and argumentsToStringFull for package built with grunt
 -   Fix to much logs in TelephonyService::getCallFromCache
+-   Fix TelephonyService Call instance creation on few methods.
 
 ## [1.63.0] - 2019-10-31
 -   Fix Options.ts the start up service requested to start when it is not already present in default config.